fix: separate custom hook
